refactor: extract test modules to sibling files (5 large prod files split)
Mechanical test-module extraction across 5 files where the test module
dominated source-file line count (>35% of file). Production code is
unchanged; only `#[cfg(test)] mod tests { ... 1000-line body ... }` is
replaced with `#[cfg(test)] mod tests;` + the body relocated to a
sibling `tests.rs` per the Rust 2018+ module-file convention.

Net production-file line counts (5 files, headers before/after):

  fragments.rs        1686 -> 458   (-73%)
  extract_helpers.rs  1577 -> 456   (-71%)
  adaptive.rs         1033 -> 555   (-46%)
  pipeline/mod.rs      954 -> 434   (-54%)
  security/lib.rs      839 -> 547   (-35%)

Total production lines moved out of view: ~3679. Tests stay in the same
crate, still discoverable via `cargo test --workspace`, and continue to
use `#[cfg(test)]` gating.

Side effects:
- pipeline/mod.rs: the original `#[allow(clippy::disallowed_methods)]`
  attribute on the inner `mod tests {}` is now applied to the `mod
  tests;` declaration.
- extract_helpers/tests.rs: `include_str!` paths gain one extra `../`
  level because the test file is now one directory deeper.
- security/tests.rs: lives as a sibling of lib.rs (Rust 2018+ rule:
  lib.rs IS the crate root; `mod tests;` resolves to a sibling file or
  `tests/mod.rs`, never `lib/tests.rs`).

No production code, no public surface, no behavior changed. The 82
test suites continue to pass. Pre-PR gate: cargo fmt --check, cargo
clippy --workspace --all-targets -- -D warnings, cargo test
--workspace, bash scripts/check-dedup.sh — all green.

Tiered next targets (deferred to a separate PR with real structural
work, not pure test extraction):
- tnaflix/mod.rs (1342) — 4 extractor structs in one file; warrants
  per-extractor file split.
- commands/download.rs (987), koreanpornmovie/mod.rs (1004),
  client/mod.rs (770), video_convert.rs (928) — survey for split
  candidates.
